# Make loading of pretrained gpt2 faster by avoiding initialization of Conv1D's weights